notebooks/repo-scraping/requirements.txt
pandas
requests
pyarrow
fastparquet
python-dotenv
